[IMP] mail: unskip "thread notifications are re-ordered on receiving a new message"
